+++ src/Render/Mesh.cpp
@@ -4,11 +4,11 @@
 #include <WICTextureLoader.h>
 
 CMesh::CMesh(ID3D11Device* device, std::vector<MeshVertex> vertices, std::vector<unsigned int> indices, ID3D11ShaderResourceView* texture)
-    : NumIndices(indices.size()), Texture(texture)
+    : NumIndices(static_cast<unsigned int>(indices.size())), Texture(texture)
 {
     D3D11_BUFFER_DESC vbd;
     vbd.Usage = D3D11_USAGE_IMMUTABLE;
-    vbd.ByteWidth = sizeof(MeshVertex) * vertices.size();
+    vbd.ByteWidth = sizeof(MeshVertex) * static_cast<UINT>(vertices.size());
     vbd.BindFlags = D3D11_BIND_VERTEX_BUFFER;
     vbd.CPUAccessFlags = 0;
     vbd.MiscFlags = 0;
@@ -20,7 +20,7 @@ CMesh::CMesh(ID3D11Device* device, std::vector<MeshVertex> vertices, std::vector
 
     D3D11_BUFFER_DESC ibd;
     ibd.Usage = D3D11_USAGE_IMMUTABLE;
-    ibd.ByteWidth = sizeof(UINT) * indices.size();
+    ibd.ByteWidth = sizeof(UINT) * static_cast<UINT>(indices.size());
     ibd.BindFlags = D3D11_BIND_INDEX_BUFFER;
     ibd.CPUAccessFlags = 0;
     ibd.MiscFlags = 0;
